% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aria2c.R
\name{aria2c}
\alias{aria2c}
\alias{aria2c_finished}
\title{aria2c}
\usage{
aria2c(
  infile,
  outdir = "OUTPUT",
  j = 1,
  x = 4,
  s = 5,
  verbose = TRUE,
  run = TRUE,
  options = ""
)

aria2c_finished(indir, subfix = "*.nc$|*.nc4$", subfix_temp = "*.aria2$")
}
\arguments{
\item{j}{--max-concurrent-downloads=N Set maximum number of parallel downloads for
every static (HTTP/FTP) URL, torrent and metalink.
See also --split and --optimize-concurrent-downloads options.
Possible Values: 1-*, Default: 5, Tags: #basic}

\item{x}{--max-connection-per-server=NUM The maximum number of connections to one
server for each download.
Possible Values: 1-16, Default: 1, Tags: #basic, #http, #ftp}

\item{s}{--split=N Download a file using N connections. If more than N URIs
are given, first N URIs are used and remaining URLs are used for backup.
If less than N URIs are given, those URLs are used more than once so that
N connections total are made simultaneously. The number of connections to the
same host is restricted by the --max-connection-per-server option. See also the
--min-split-size option.
Possible Values: 1-*, Default: 5, Tags: #basic, #http, #ftp}

\item{verbose}{whether to print the downloading command line?}

\item{options}{other parameters to \code{aria2c}}
}
\description{
aria2c
}
