% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seach_esgf.R
\name{build_esgf_url}
\alias{build_esgf_url}
\title{build_esgf_url}
\usage{
build_esgf_url(
  ...,
  param = NULL,
  host = "https://esgf-node.llnl.gov/esg-search/search"
)
}
\arguments{
\item{...}{named parameters}

\item{param}{list parameters}
}
\description{
build_esgf_url
}
