% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dfile.R
\name{check_dfile}
\alias{check_dfile}
\title{make sure no overlaping in date}
\usage{
check_dfile(d_file, verbose = 0)
}
\arguments{
\item{d_file}{object returned by \code{\link[=CMIP5Files_summary]{CMIP5Files_summary()}}}

\item{verbose}{echo duplicated info if \code{verbose >= 2}}
}
\description{
make sure no overlaping in date
}
\keyword{internal}
