% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIPFiles.R, R/CMIPFiles_filter.R
\name{get_ensemble}
\alias{get_ensemble}
\alias{extract_model}
\alias{CMIP5Files_info}
\alias{CMIP5Files_summary}
\alias{CMIP5Files_filter}
\title{summary CMIP5 files information}
\usage{
get_ensemble(files)

extract_model(files)

CMIP5Files_info(files)

CMIP5Files_summary(d)

CMIP5Files_filter(
  files,
  duration = 200,
  period = NULL,
  check_dupli = TRUE,
  verbose = 1
)
}
\arguments{
\item{files}{CMIP5 nc files, full name path.}

\item{d}{An object returned by \code{CMIP5Files_filter} or
\code{CMIP5Files_info} or a data.frame at least with the columns of \verb{ 'model', 'year_start_adj', 'year_end', 'file'}.}

\item{duration}{length of year.}

\item{period}{starting and ending year, e.g. \code{c(1850, 2005)}. If
provided, duration will be overwritten.
If duration and period are \code{null} at the same time, no filter applied.}

\item{check_dupli}{Boolean. If true, duplicated date will be removed by \code{\link[=check_dfile]{check_dfile()}}.}

\item{verbose}{\itemize{
\item echo nc file missing info, if \code{verbose >= 1}
\item echo duplicated info, if \code{verbose >= 2}
}}
}
\value{
d_files A data.table with colnames of
\verb{'Id', 'model', 'kind', 'start', 'end', 'year_start', 'year_start_adj', 'year_end', 'file'}
}
\description{
\itemize{
\item \link{get_ensemble}: Extract CMIP5 ensemle name from file names
\item \link{extract_model}: Extract CMIP5 model name from file names
\item \link{CMIP5Files_info}: Extract CMIP5 information from file names
\item \link{CMIP5Files_filter}: Filter corresponding duration or period files.
Dates has been adjust from 12-01 to 01-01 for \code{start_adj}. Duplicated dates
are also removed.
\item \link{CMIP5Files_summary}: Get the start and end information of every model
}
}
\note{
Only one scenario per time.
}
\examples{
data(files_short)
ensemble <- get_ensemble(files_short)
model <- extract_model(files_short)
CMIP5Files_info(files_short)

# Get the summary infomation of CMIP5Files_filter or CMIP5Files_info
CMIP5Files_filter(files_short) \%>\% CMIP5Files_summary()

# filter CMIP5 files
CMIP5Files_filter(files_short, duration = 200)

}
\seealso{
\code{\link[=check_dfile]{check_dfile()}}, \code{\link[=rm_duplicate]{rm_duplicate()}}
}
