% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIPFiles.R
\name{get_varname}
\alias{get_varname}
\alias{get_CMIP5}
\alias{get_scenario}
\alias{get_scenario2}
\alias{get_model}
\alias{get_freq}
\alias{get_period}
\title{get CMIP5 scenario and model from file path}
\usage{
get_varname(file)

get_scenario(file)

get_scenario2(file)

get_model(file, prefix = "_", postfix = "_")

get_freq(file)

get_period(scenario = "historical", period = NULL)
}
\description{
get CMIP5 scenario and model from file path
}
\keyword{NULL}
\keyword{internal}
