% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_terminal.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bold}
\alias{red}
\alias{green}
\alias{underline}
\alias{\%>\%}
\alias{ping}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{crayon}{\code{\link[crayon:crayon]{bold}}, \code{\link[crayon:crayon]{green}}, \code{\link[crayon:crayon]{red}}, \code{\link[crayon:crayon]{underline}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{pingr}{\code{\link[pingr]{ping}}}
}}

