% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dfile.R
\name{rm_duplicate}
\alias{rm_duplicate}
\title{Remove duplicated date in \code{CMIP5Files_summary}}
\usage{
rm_duplicate(d, verbose = 0)
}
\arguments{
\item{d}{An object returned by \code{CMIP5Files_filter} or
\code{CMIP5Files_info} or a data.frame at least with the columns of \verb{ 'model', 'year_start_adj', 'year_end', 'file'}.}

\item{verbose}{echo duplicated info if \code{verbose >= 2}}
}
\description{
Currently only \code{HadGEM2} have overlaping time-period in ncfiles. This
function works well for this situation.
}
\keyword{internal}
