% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seach_esgf.R
\name{build_esgf_param}
\alias{build_esgf_param}
\alias{build_esgf_url}
\alias{retrieve_esgf_docs}
\alias{search_esgf}
\title{search_esgf}
\usage{
build_esgf_param(..., param = NULL)

build_esgf_url(
  ...,
  param = NULL,
  host = "https://esgf-node.llnl.gov/esg-search/search"
)

retrieve_esgf_docs(
  variable_id = "tasmax",
  table_id = "day",
  experiment_id = "historical",
  member_id = "r1i1p1f1",
  source_id = NULL,
  ...,
  param = NULL
)

search_esgf(
  variable_id = "tasmax",
  table_id = "day",
  experiment_id = "historical",
  member_id = "r1i1p1f1",
  source_id = NULL,
  ...,
  url_type = c("HTTPServer", "OPENDAP"),
  param
)
}
\arguments{
\item{...}{named parameters}

\item{param}{list parameters}
}
\description{
search_esgf
}
\details{
\subsection{Parameters}{
\itemize{
\item project       : "CMIP6"
\item type          : "File"
\item table_id      : "day"
\item variable_id   : "tasmax"
\item experiment_id : "historical"
\item source_id     : "ACCESS-CM2"
\item member_id     : "r1i1p1f1"
\item latest        : TRUE
}
}
}
